Move collectJarsFromClasspathFile in its own module
