+++ java-checks-aws/pom.xml
@@ -45,6 +45,12 @@
     </dependency>
 
     <dependency>
+      <groupId>org.sonarsource.java</groupId>
+      <artifactId>test-classpath-reader</artifactId>
+      <version>${project.version}</version>
+      <scope>test</scope>
+    </dependency>
+    <dependency>
       <groupId>${project.groupId}</groupId>
       <artifactId>java-checks-testkit</artifactId>
       <version>${project.version}</version>
