+++ java-checks-test-sources/test-classpath-reader/src/test/resources/classpath-example.txt
