+++ java-checks-test-sources/test-classpath-reader/src/test/resources/invalid-classpath-example.txt
