+++ sonar-java-plugin/pom.xml
@@ -123,6 +123,12 @@
       <scope>test</scope>
     </dependency>
     <dependency>
+      <groupId>org.sonarsource.java</groupId>
+      <artifactId>test-classpath-reader</artifactId>
+      <version>${project.version}</version>
+      <scope>test</scope>
+    </dependency>
+    <dependency>
       <groupId>${project.groupId}</groupId>
       <artifactId>java-checks-testkit</artifactId>
       <version>${project.version}</version>
